Test of Equality of Covariances given by Schott 2001

Performs 2 and k sample equality of covariance matrix test using Schott 2001

Schott2001(x, ...)

Arguments

x
data as data.frame, grouped_df, resample or matrix object

...
other options passed to functions

Value

Test statistic of the hypothesis test

Examples

irisSpecies <- unique(iris$Species)

iris_ls <- lapply(irisSpecies,
    function(x){as.matrix(iris[iris$Species == x, 1:4])}
                 )

names(iris_ls) <- irisSpecies

Chaipitak2013(iris_ls)

#> 
#> 	Chaipitak and Chongchareon 2013 Equality of Covariance Test
#> 
#> data:  setosa, versicolor and virginica
#> Chi Squared = 3.8455, df = 2, p-value = 0.1462
#> alternative hypothesis: true difference in covariances is not equal to 0
#>