Test of Structure of a Covariance Matrix given by Srivastava 2005

Test of Structure of a Covariance Matrix given by Srivastava 2005

Srivastava2005(x, Sigma = "identity", ...)

Arguments

x
data

Sigma
Population covariance matrix

...
other options passed to covTest method

Value

A list with class "htest" containing the following components:

statistic 	 the value of equality of covariance test statistic

parameter 	 the degrees of freedom for the chi-squared statistic

p.value 	 the p=value for the test

estimate 	 the estimated covariances if less than 5 dimensions

null.value 	 the specified hypothesized value of the covariance difference

alternative 	 a character string describing the alternative hyposthesis

method 	 a character string indicating what type of equality of covariance test was performed

Details

The structureCovariances function is a wrapper function that formats the data for the specific covTest functions.

References

Srivastava, M. S. (2005). Some Tests Concerning the Covariance Matrix in High Dimensional Data. Journal of the Japan Statistical Society, 35(2), 251-272. 10.14490/jjss.35.251

Examples

Srivastava2005(as.matrix(iris[1:50, 1:4]))

#> 
#> 	Srivastava 2005 Test of Covariance Matrix Structure
#> 
#> data:  
#> Standard Normal = 21.479, Mean = 0, Variance = 1, p-value < 2.2e-16
#> alternative hypothesis: true difference between the Sample Covariance Matrix and the Null Covariance Matrix Structure is not equal to 0
#> sample estimates:
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   0.12424898 0.099216327  0.016355102 0.010330612
#> Sepal.Width    0.09921633 0.143689796  0.011697959 0.009297959
#> Petal.Length   0.01635510 0.011697959  0.030159184 0.006069388
#> Petal.Width    0.01033061 0.009297959  0.006069388 0.011106122
#>