Tests for Homogeneity of Covariance Matrices

Performs tests for homogeneity of 2 and k covariance matrices.

Ahmad2017(x, ...)

BoxesM(x, ...)

Chaipitak2013(x, ...)

Ishii2016(x, ...)

Schott2001(x, ...)

Schott2007(x, ...)

Srivastava2007(x, ...)

Srivastava2014(x, ...)

SrivastavaYanagihara2010(x, ...)

Arguments

x
data as a list of matrices

...
other options passed to covTest method

Value

A list with class "htest" containing the following components:

statistic 	 the value of homogeneity of covariance test statistic

parameter 	 the degrees of freedom for the chi-squared statistic

p.value 	 the p=value for the test

estimate 	 the estimated covariances if less than 5 dimensions

null.value 	 the specified hypothesized value of the covariance difference

alternative 	 a character string describing the alternative hyposthesis

method 	 a character string indicating what type of homogeneity of covariance test was performed

References

Ahmad, R. (2017). Location-invariant test of homogeneity of large-dimensional covariance matrices. Journal of Statistical Theory and Practice, 11(4):731-745. 10.1080/15598608.2017.1308895

Chaipitak, S. and Chongcharoen, S. (2013). A test for testing the equality of two covariance matrices for high-dimensional data. Journal of Applied Sciences, 13(2):270-277. 10.3923/jas.2013.270.277

Ishii, A., Yata, K., and Aoshima, M. (2016). Asymptotic properties of the first pricipal component and equality tests of covariance matrices in high-dimesion, low-sample-size context. Journal of Statistical Planning and Inference, 170:186-199. 10.1016/j.jspi.2015.10.007

Schott, J (2001). Some Tests for the Equality of Covariance Matrices. Journal of Statistical Planniing and Inference. 94(1), 25-36. 10.1016/S0378-3758(00)00209-3

Schott, J. (2007). A test for the equality of covariance matrices when the dimension is large relative to the sample sizes. Computational Statistics & Data Analysis, 51(12):6535-6542. 10.1016/j.csda.2007.03.004

Srivastava, M. S. (2007). Testing the equality of two covariance matrices and independence of two sub-vectors with fewer observations than the dimension. InInternational Conference on Advances in InterdisciplinaryStistics and Combinatorics, University of North Carolina at Greensboro, NC, USA.

Srivastava, M., Yanagihara, H., and Kubokawa T. (2014). Tests for covariance matrices in high dimension with less sample size. Journal of Multivariate Analysis, 130:289-309. 10.1016/j.jmva.2014.06.003

Srivastava, M. and Yanagihara, H. (2010). Testing the equality of several covariance matrices with fewer observation that the dimension. Journal of Multivariate Analysis, 101(6):1319-1329. 10.1016/j.jmva.2009.12.010

See also

Other Testing for Homogeneity of Covariance Matrices: homogeneityCovariances

Examples

irisSpecies <- unique(iris$Species)

iris_ls <- lapply(irisSpecies,
    function(x){as.matrix(iris[iris$Species == x, 1:4])}
                 )

names(iris_ls) <- irisSpecies

Ahmad2017(iris_ls)

#> 
#> 	Ahmad 2017 Homogeneity of Covariance Matrices Test
#> 
#> data:  setosa, versicolor and virginica
#> Standard Normal = 1.1193, Mean = 0, Variance = 1, p-value = 0.1315
#> alternative hypothesis: true difference in covariance matrices is not equal to 0
#>